Reviewer 1 Peer assessment of assignment 3
This would run on titanium, but I found an error on line 3, 65, and 88. The background color 'fff' needs a '#' symbol before it, on line 65 you included a comma at the end of of the bracket, and on line 88 you're missing a semicolon, but once those were corrected I was able to run it just fine. I was able to use the keys you listed to produce the different texts you posted and you used 10 colors too. I can't run the navigation buttons (I tried clicking it and re-running the code) for some reason even though the program is running, but I don't see any errors and titanium isn't reporting anything in my compiler so I have no idea how to fix those. You've included the camera gadget from page 154, chapter 7,and I can interact with it so it seems to be fine. Overall the code's not bad, but the only major issue is with the navigation windows since they don't work for some reason.